Move utilties out of the `WalletList` component
These are used everywhere, and deserve their own file.
